first pass at a synth.py
