+++ tulip/shared/py/tulip_queue.py
@@ -1,0 +1,77 @@
+# Micropython collections.deque does not support remove.
+class Queue:
+    def __init__(self, maxsize=64, name=''):
+        self.name = name    # Just for debug.
+        self.maxsize = maxsize + 1
+        self.queue = [None] * self.maxsize
+        self.head = 0
+        self.tail = 0
+
+    def _next(self, pointer):
+        """Incrementing a cicular buffer pointer."""
+        return (pointer + 1) % self.maxsize
+                
+    def _prev(self, pointer):
+        """Decrementing a cicular buffer pointer."""
+        return (pointer + self.maxsize - 1) % self.maxsize
+                
+    def put(self, item):
+        self.queue[self.tail] = item
+        self.tail = self._next(self.tail)
+        if self.tail == self.head:
+            # Wrap around
+            self.head = self._next(self.head)
+            print("%s: dropped oldest item", self)
+
+    def _delete_at(self, pointer):
+        """Remove the value at queue[pointer], and close up the rest."""
+        if self.tail > pointer:
+            self.queue[pointer : self.tail - 1] = (
+                self.queue[pointer + 1 : self.tail])
+            self.tail = self._prev(self.tail)
+        elif self.tail < pointer:
+            # Shift head: pointer one step forward.
+            self.queue[self.head + 1 : pointer + 1] = self.queue[self.head : pointer]
+            self.head = self._next(self.head)
+        else:
+            raise ValueError('pointer at tail???')
+
+    def remove(self, value):
+        """Remove first occurrence of value from queue."""
+        pointer = self.head
+        while pointer != self.tail:
+            if self.queue[pointer] == value:
+                self._delete_at(pointer)
+                return
+            pointer = self._next(pointer)
+        # Fell through, value wasn't found.
+        raise ValueError('%s: No such item: %s' % (self, str(value)))
+
+    def empty(self):
+        return self.head == self.tail
+
+    def full(self):
+        return self.head == self._next(self.tail)
+
+    def qsize(self):
+        return (self.tail - self.head + self.maxsize) % self.maxsize
+
+    def get(self):
+        if self.empty():
+            # get() on empty queue.
+            raise ValueError
+        value = self.queue[self.head]
+        self.head = self._next(self.head)
+        return value
+
+    def __repr__(self):
+        result = []
+        p = self.head
+        while p != self.tail:
+            result.append(self.queue[p])
+            p = self._next(p)
+        return ("Queue%s(maxsize=%d) [" % (self.name, self.maxsize - 1)
+            + (", ".join(str(s) for s in result))
+            + "]")
+
+
